Update BMDN and simplify test invocation
