+++ tests/Our.Umbraco.Ditto.PerformanceTests/packages.config
@@ -1,4 +1,23 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="BenchmarkDotNet" version="0.9.4" targetFramework="net45" />
+  <package id="BenchmarkDotNet" version="0.10.3" targetFramework="net45" />
+  <package id="BenchmarkDotNet.Core" version="0.10.3" targetFramework="net45" />
+  <package id="BenchmarkDotNet.Diagnostics.Windows" version="0.10.3" targetFramework="net45" />
+  <package id="BenchmarkDotNet.Toolchains.Roslyn" version="0.10.3" targetFramework="net45" />
+  <package id="Microsoft.CodeAnalysis.Analyzers" version="1.1.0" targetFramework="net45" />
+  <package id="Microsoft.CodeAnalysis.Common" version="1.3.2" targetFramework="net45" />
+  <package id="Microsoft.CodeAnalysis.CSharp" version="1.3.2" targetFramework="net45" />
+  <package id="Microsoft.Diagnostics.Tracing.TraceEvent" version="1.0.41" targetFramework="net45" />
+  <package id="System.Collections" version="4.0.0" targetFramework="net45" />
+  <package id="System.Collections.Immutable" version="1.1.37" targetFramework="net45" />
+  <package id="System.Diagnostics.Debug" version="4.0.0" targetFramework="net45" />
+  <package id="System.Globalization" version="4.0.0" targetFramework="net45" />
+  <package id="System.Linq" version="4.0.0" targetFramework="net45" />
+  <package id="System.Reflection.Metadata" version="1.2.0" targetFramework="net45" />
+  <package id="System.Resources.ResourceManager" version="4.0.0" targetFramework="net45" />
+  <package id="System.Runtime" version="4.0.0" targetFramework="net45" />
+  <package id="System.Runtime.Extensions" version="4.0.0" targetFramework="net45" />
+  <package id="System.Threading" version="4.0.0" targetFramework="net45" />
+  <package id="System.Threading.Tasks" version="4.0.11" targetFramework="net45" />
+  <package id="System.Threading.Tasks.Extensions" version="4.3.0" targetFramework="net45" />
 </packages>
